+++ packages/article-byline/__tests__/web/article-byline.web.test.js
@@ -1,7 +1,6 @@
+import ArticleByline from "../../src/article-byline";
 import shared from "../shared";
-import sharedBylineWithLinks from "../shared-byline-with-links";
 
-describe("Article Byline test on web", () => {
-  shared();
-  sharedBylineWithLinks();
+describe("ArticleByline tests on web", () => {
+  shared(ArticleByline);
 });