+++ packages/article-byline/src/article-byline-with-links.js
@@ -1,10 +1,7 @@
 import React from "react";
 import { TextLink } from "@times-components/link";
 import { getBylineStyles, renderByline } from "./utils";
-import {
-  articleBylinePropTypes,
-  articleBylineDefaultPropTypes
-} from "./article-byline-prop-types";
+import { propTypes, defaultProps } from "./article-byline-prop-types";
 
 const renderAuthorComponent = (
   children,
@@ -30,8 +27,9 @@ const ArticleBylineWithLinks = ({ ast, style, color, onAuthorPress }) => {
   const bylineStyles = getBylineStyles(style, color);
   return renderByline(renderAuthorComponent, ast, bylineStyles, onAuthorPress);
 };
+ArticleBylineWithLinks.displayName = "ArticleBylineWithLinks";
 
-ArticleBylineWithLinks.propTypes = articleBylinePropTypes;
-ArticleBylineWithLinks.defaultProps = articleBylineDefaultPropTypes;
+ArticleBylineWithLinks.propTypes = propTypes;
+ArticleBylineWithLinks.defaultProps = defaultProps;
 
 export default ArticleBylineWithLinks;