+++ packages/button/src/button-prop-types.js
@@ -4,13 +4,12 @@ import PropTypes from "prop-types";
 const { style: ViewPropTypesStyle } = ViewPropTypes;
 
 export const propTypes = {
-  onPress: PropTypes.func,
+  onPress: PropTypes.func.isRequired,
   style: ViewPropTypesStyle,
   title: PropTypes.string
 };
 
 export const defaultProps = {
-  onPress: () => {},
   style: null,
   title: "Submit"
 };