+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.18.4"></a>
+## [0.18.4](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.18.3...@times-components/article-byline@0.18.4) (2018-06-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.18.3"></a>
+## [0.18.3](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.18.2...@times-components/article-byline@0.18.3) (2018-06-12)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
 <a name="0.18.2"></a>
 ## [0.18.2](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.18.1...@times-components/article-byline@0.18.2) (2018-06-08)