+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.1"></a>
+## [0.11.1](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.11.0...@times-components/styleguide@0.11.1) (2018-06-13)
+
+
+
+
+**Note:** Version bump only for package @times-components/styleguide
+
 <a name="0.11.0"></a>
 # [0.11.0](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.10.1...@times-components/styleguide@0.11.0) (2018-06-07)